+++ tests/unit/server/services/storage/index.test.ts
@@ -82,7 +82,7 @@ describe('Storage Service', () => {
           size: 0,
           aggs: {
             test_cases: {
-              terms: { field: 'id', size: 10000 },
+              terms: { field: 'id.keyword', size: 10000 },
               aggs: {
                 latest: {
                   top_hits: { size: 1, sort: [{ version: { order: 'desc' } }] },